Rename authMiddleware to authenticationMiddleware
