+++ src/api/routers/authRouter.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 
-import { authMiddleware } from "../middlewares/authMiddleware.js";
+import { authMiddleware } from "../middlewares/authenticationMiddleware.js";
 import AuthCotroller from "../controllers/authController.js";
 
 class AuthRouter {
@@ -12,17 +12,13 @@ class AuthRouter {
   }
 
   private initializeRoutes() {
-    this.router.post("/signup", this.authController.signupController);
-    this.router.post("/login", this.authController.loginController);
-    this.router.post(
-      "/logout",
-      authMiddleware,
-      this.authController.logoutController
-    );
+    this.router.post("/signup", this.authController.signup);
+    this.router.post("/login", this.authController.login);
+    this.router.post("/logout", authMiddleware, this.authController.logout);
     this.router.post(
       "/refresh-token",
       authMiddleware,
-      this.authController.refreshController
+      this.authController.refresh
     );
   }
